sua giao dien

+++ src/main/webapp/product/home.jsp
@@ -96,9 +96,16 @@ Lower Header Section
                         <c:if test="${sessionScope.acc != null && sessionScope.acc.roleId == 1}">
                             <li class=""><a href="/products?action=create">Thêm sản phẩm mới</a></li>
                         </c:if>
-                        <li class=""><a href="?action=huong-dan-mua-hang">Hướng dẫn mua hàng</a></li>
-                        <li class=""><a href="?action=gioi-thieu">Giới thiệu</a></li>
-                        <li class=""><a href="general.html">Tin tức</a></li>
+                        <c:if test="${sessionScope.acc != null && sessionScope.acc.roleId == 1}">
+                            <li class=""><a href="/products?action=addCategory">Thêm danh mục</a></li>
+                        </c:if>
+                        <c:if test="${sessionScope.acc != null && sessionScope.acc.roleId == 1}">
+                            <li class=""><a href="/products?action=addBrand">Thêm nhãn hiệu</a></li>
+                        </c:if>
+                        <c:if test="${sessionScope.acc == null || sessionScope.acc.roleId == 2}">
+                            <li class=""><a href="/?action=huong-dan-mua-hang">Hướng dẫn mua hàng</a></li>
+                            <li class=""><a href="/?action=gioi-thieu">Giới thiệu</a></li>
+                        </c:if>
                     </ul>
                     <form method="get" action="/" class="navbar-search pull-right">
                         <input type="text" placeholder="Search" class="search-query span2" name="Search">
@@ -265,7 +272,7 @@ Lower Header Section
                                                    title="add to cart"> Add to cart</a></h4>
                                         </c:if>
                                         <c:if test="${sessionScope.acc != null && sessionScope.acc.roleId == 1}">
-                                            <h4><a class="shopBtn" href="/products?action=delete&id=${product.id}"
+                                            <h4><a onclick="return confirm('Are you sure?')" class="shopBtn" href="/products?action=delete&id=${product.id}"
                                                    title="add to cart"> Delete</a></h4>
                                         </c:if>
                                         <c:if test="${sessionScope.acc != null && sessionScope.acc.roleId == 1}">
